Add Feature Face Recognition
The functions will be triggered by Front-end through SignalR.
Function SendFaceVerifyUrl uses two urls and return the confidence of similarity between the two pictures.
Function SendFaceCompareName use one image in the designated folder and compare with another folder with lots of images inside. It will return the picutre name with the highest confidence of similarity.
The picture can't be bigger than 32k if uploaded from local.